pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

# ==============================================================================
# Métadonnées du projet
# ==============================================================================

[project]
name = "Prométhée-ai"
version = "8.0.0"
description = "Assistant IA desktop souverain avec outils intégrés, RAG et support Légifrance"
readme = "README.md"
license = { text = "Proprietary" }
requires-python = ">=3.10"

keywords = ["ai", "llm", "rag", "desktop", "assistant", "legifrance", "pyqt6"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# ── Dépendances obligatoires ──────────────────────────────────────────────────
dependencies = [
    "python-dotenv>=1.0.0",
    "openai>=1.30.0",
    "PyQt6>=6.6.0",
    "PyQt6-WebEngine>=6.6.0",
    "PyYAML>=6.0.1",
    "requests>=2.31.0",
    "httpx>=0.27.0",
]

# ── Dépendances optionnelles (groupes) ────────────────────────────────────────
[project.optional-dependencies]

# Outils web
web = [
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.12.0",
    "lxml>=5.0.0",
]

# Analyse de données
data = [
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",
    "numpy>=1.26.0",
]

# RAG — base vectorielle
rag = [
    "qdrant-client>=1.9.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
]

# PDF et OCR
ocr = [
    "PyMuPDF>=1.24.0",
    "Pillow>=10.3.0",
    "pytesseract>=0.3.10",
]

# Documents Office
office = [
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.0",
]

# Audio
audio = [
    "pydub>=0.25.1",
]

# Modèles locaux Ollama
local = [
    "ollama>=0.2.0",
]

# SQL (drivers additionnels — installés à la demande via python_install)
sql = [
    "psycopg2-binary>=2.9.0",   # PostgreSQL
    "pymysql>=1.1.0",            # MySQL / MariaDB
]

# Tout (sauf drivers SQL)
all = [
    "promethee-ai[web,data,rag,ocr,office,audio,local]",
]

# Développement
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "coverage>=7.5.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.12",
]

# ── Point d'entrée ────────────────────────────────────────────────────────────
[project.scripts]
promethee = "main:run"

[project.gui-scripts]
promothee-gui = "main:run"

# ==============================================================================
# Outils
# ==============================================================================

# ── Ruff (linter + formateur) ─────────────────────────────────────────────────
[tool.ruff]
target-version = "py310"
line-length = 100
src = [".", "core", "tools", "ui"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (géré par le formateur)
    "B008",   # do not perform function calls in default arguments
    "SIM108", # ternary operator (parfois moins lisible)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S102"]  # assert et exec autorisés dans les tests

# ── Mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true   # PyQt6 stubs non exhaustifs
exclude = ["tests/", "ui/"]    # UI PyQt6 difficile à typer statiquement

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

# ── Coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["core", "tools"]
omit = [
    "tests/*",
    "ui/*",
    "*/site-packages/*",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

# ── Setuptools ────────────────────────────────────────────────────────────────
[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "tools*", "ui*"]
exclude = ["tests*", "assets*"]
